rename: ConceptSlug → VideoDiagram
And move Position types into core: [sys.types] module
