+++ code/system.ui/react.common/src/ui/Position/types.mts
@@ -1,16 +1,6 @@
 import { type t } from './common';
 
 /**
- *  Coordinate Positioning
- * (spacial placement on 2D rectangular plane)
- */
-export type PositionX = 'left' | 'right' | 'center';
-export type PositionY = 'top' | 'bottom' | 'center';
-export type Position = { x: PositionX; y: PositionY };
-export type Pos = [PositionX, PositionY];
-export type PositionInput = Position | Pos;
-
-/**
  * Component
  */
 export type PositionProps = {
