+++ code/system.ui/react.concept/src/Concept.ts
@@ -1,0 +1,7 @@
+import { PlayBar } from './ui/ui.PlayBar';
+import { VideoDiagram } from './ui/ui.VideoDiagram';
+
+export const Concept = {
+  PlayBar,
+  VideoDiagram,
+} as const;
