+++ code/system.ui/react.concept/src/common/types.mts
@@ -7,7 +7,15 @@ export type { Observable } from 'rxjs';
  * @system
  */
 export type { SpecImport, TestSuiteRunResponse } from 'sys.test.spec/src/types.mjs';
-export type { Disposable, EventBus, Lifecycle } from 'sys.types/src/types.mjs';
+export type {
+  Disposable,
+  EventBus,
+  Lifecycle,
+  Pos,
+  Position,
+  PositionInput,
+} from 'sys.types/src/types.mjs';
+export type { ImageSizeStrategy } from 'sys.ui.react.media.image/src/types.mjs';
 
 /**
  * @local
