+++ code/system.ui/react.concept/src/test.ui/entry.Specs.mts
@@ -3,7 +3,7 @@ export { Pkg } from '../index.pkg.mjs';
 export const Specs = {
   'sys.ui.concept.tests': () => import('./-TestRunner'),
   'sys.ui.concept.ui.Info': () => import('../ui/Info/-dev/-SPEC'),
-  'sys.ui.concept.ui.ConceptSlug': () => import('../ui/Concept.Slug/-SPEC'),
+  'sys.ui.concept.ui.VideoDiagram': () => import('../ui/ui.VideoDiagram/-SPEC'),
   'sys.ui.concept.ui.PlayBar': () => import('../ui/ui.PlayBar/-SPEC'),
 };
