--- code/system.ui/react.concept/src/ui/Concept/index.ts
@@ -1,7 +1,0 @@
-import { PlayBar } from '../ui.PlayBar';
-import { ConceptSlug } from '../Concept.Slug';
-
-export const Concept = {
-  PlayBar: PlayBar,
-  Slug: ConceptSlug,
-} as const;
