+++ code/system.ui/react.concept/src/ui/ui.VideoDiagram/common.ts
