+++ code/system.ui/react.concept/src/ui/ui.VideoDiagram/types.mts
@@ -1,0 +1,10 @@
+import { type t } from './common';
+
+/**
+ * Component
+ */
+export type VideoDiagramProps = {
+  vimeo?: t.VimeoInstance;
+  slug?: t.ConceptSlug;
+  style?: t.CssValue;
+};
