+++ code/system.ui/react.concept/src/ui/ui.VideoDiagram/ui.Dummy.tsx
