+++ code/system.ui/react.concept/src/ui/ui.VideoDiagram/ui.Image.tsx
